skeleton for leverage page
