+++ packages/ui/components/pages/LeveragePageOld/LeverageList/NewPosition/AdditionalInfo/CreatePositionButton/CreatePositionModal/PendingTransaction.tsx
