+++ packages/ui/components/pages/LeveragePageOld/LeverageList/OpenPosition/AdditionalInfo/AdjustRatioButton/AdjustRatioModal/LeverageSlider.tsx
