+++ packages/ui/components/pages/LeveragePageOld/LeverageList/OpenPosition/AdditionalInfo/ClosePositionButton/ClosePositionModal/PendingTransaction.tsx
