+++ packages/ui/components/pages/LeveragePageOld/LeverageList/OpenPosition/DebtValue.tsx
