+++ packages/ui/components/pages/LeveragePageOld/LeverageList/OpenPosition/PositionValue.tsx
