+++ packages/ui/components/pages/LeveragePageOld/LeverageList/OpenPosition/SafetyBuffer.tsx
